MLE-15791 Refactored aggregate XML classes into a package
Had to make some stuff public, but this makes it much easier to see all the aggregate XML code in one place. Will make performance improvement next.
